finalize count recs mr job

+++ mr_countrecs/src/CountRecsMapper.java
@@ -1,4 +1,3 @@
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -6,10 +5,10 @@ import org.apache.hadoop.mapreduce.Mapper;
 import java.io.IOException;
 
 public class CountRecsMapper
-  extends Mapper<LongWritable, Text, Text, IntWritable> {
+  extends Mapper<LongWritable, Text, Text, LongWritable> {
 
   Text keyText = new Text("");
-  private static final IntWritable one = new IntWritable(1);
+  private static final LongWritable one = new LongWritable(1);
 
   @Override
   public void map(LongWritable key, Text value, Context context)
