+++ mr_countrecs/src/CountRecsReducer.java
@@ -1,4 +1,3 @@
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -6,15 +5,15 @@ import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
 
 public class CountRecsReducer
-  extends Reducer<Text, IntWritable, Text, LongWritable> {
+  extends Reducer<Text, LongWritable, Text, LongWritable> {
   
   @Override
-  public void reduce(Text key, Iterable<IntWritable> values,
+  public void reduce(Text key, Iterable<LongWritable> values,
       Context context)
       throws IOException, InterruptedException {
 
     long sum = 0;
-    for (IntWritable val : values) {
+    for (LongWritable val : values) {
       sum += val.get();
     }
